Fix for build error
